requirements.txt
# Python Version 3.9

llama-index
pypdf
python-dotenv
pandas
spotipy
SpeechRecognition
Selenium
pywhatkit
whatsapp-python
pywin32
python-box
icecream
torch
nltk
scikit-learn
spacy